Begin switching payment engine to 3cities
This commit begins the processs of payments being submitted and verified
using the 3cities decentralized payment processor instead of by this
plugin itself.

Goals of this switch include
- prevent pretix-eth from needing to own any code that touches onchain.
  All onchain interactions are now owned by 3cities in the 3cities
  codebase.
- enable users to pay using a variety of tokens and L2s, with the
  specific token and chain used determined at the moment the user
  submits their payment transaction.

Missing components as of this initial commit
- the 3cities payment verifier is currently a stub
- the 3cities client and verifier are expected to be running on
  localhost and can't yet be fully configured in the plugin settings
- payments must currently be in USD. Both pretix-eth and 3cities support
  ETH payments, but the glue work for pretix to pass 3cities an offered
  ETHUSD exchange rate is not yet complete

Note that before adopting 3cities, a user's smart contract wallet
eip1271 signature was verified in this pretix-eth server before the user
was allowed to send their payment transaction. eip1271 signatures are
now verified in the 3cities clientside iframe and serverside verifier
instead of in pretix-eth.