Polish for production
Add new insecure and untrusted admin convenience fields to model,
control, and exporter.

Submit tx details on 3cities TransactionSigned iframe msg instead of on
Checkout iframe msg. This causes the details to be saved before
transaction confirmation, which reduces the time window and thus risk of
the race condition where the user closes the tab before tx details are
saved in pretix.

Pass plugin payment receiver address to 3cities, such that the actual
address to which funds are sent is now controlled via plugin config.